requirements.txt
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
defusedxml==0.7.1
fonttools==4.57.0
fpdf2==2.8.2
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-vision==3.10.1
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
idna==3.10
pillow==11.2.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyTelegramBotAPI==4.26.0
python-dotenv==1.1.0
requests==2.32.3
rsa==4.9.1
urllib3==2.4.0
